test(segment): migrate scrollable
