+++ core/src/components/segment/test/scrollable/segment.e2e.ts-snapshots/segment-scrollable-md-ltr-Mobile-Chrome-linux.png
