+++ core/src/components/segment/test/scrollable/segment.e2e.ts-snapshots/segment-scrollable-md-rtl-Mobile-Safari-linux.png
